memory-consistency: split analysis into standalone pass
