+++ include/PTO/Transforms/Passes.h
@@ -52,6 +52,7 @@ createPTOBufidSyncPass(const PTOBufidSyncOptions &options = {});
 // Graph-based intra-core sync solver (coexists with PTOInsertSync).
 std::unique_ptr<Pass>
 createPTOGraphSyncSolverPass(const PTOGraphSyncSolverOptions &options = {});
+std::unique_ptr<Pass> createPTOMemoryConsistencyPass();
 // Default arch is A3 unless overridden by callers.
 std::unique_ptr<Pass> createEmitPTOManualPass();
 // Explicitly select target arch for codegen.
